LayoutTests/platform/chromium/fast/events/touch/gesture/pad-gesture-fling-expected.txt

Tests basic use of GestureFlingStart

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clientX is 10
PASS event.clientY is 11
PASS event.wheelDeltaX > 5 is true
PASS event.wheelDeltaY > 5 is true
PASS event.clientX is 10
PASS event.clientY is 11
PASS event.wheelDeltaX > 5 is true
PASS event.wheelDeltaY > 5 is true
PASS actualWheelEventsOccurred is 2
PASS successfullyParsed is true

TEST COMPLETE
